[修改]CartGetModel的資料欄位型別 [修改]CartApiController的條件篩選(前端有渲染出來,但Color選染出陣列待修改)

+++ WuliKaWu/WuliKaWu/Controllers/Api/WishListApiController.cs
@@ -115,8 +115,6 @@ namespace WuliKaWu.Controllers.Api
         [HttpGet]
         public async Task<IEnumerable<WishListGetModel>> GetWishListAsync(int productId)
         {
-            //throw new NotImplementedException();
-
             var myId = User.Claims.GetMemberId();
 
             return (await _context.WishList.Include(x => x.Product)
@@ -377,7 +375,7 @@ namespace WuliKaWu.Controllers.Api
                         WishListId = x.WishListId,
                         ProductName = x.Product.ProductName,
                         Price = x.Product.Price,
-                        SellingPrice = (decimal)x.Product.SellingPrice,
+                        SellingPrice = x.Product.SellingPrice,
                         PicturePath = x.Product.Pictures.Select(x => x.PicturePath).ToList()
                     });
                 return await wish.ToListAsync();
@@ -386,35 +384,6 @@ namespace WuliKaWu.Controllers.Api
             {
                 throw;
             }
-
-            //return (await _context.WishList
-            //        .Where(x => x.MemberId == myId)
-            //        .Select(x => new WishListGetModel
-            //        {
-            //            ProductId = productId,
-            //            MemberId = myId,
-            //            WishListId = x.WishListId,
-            //            ProductName = x.Product.ProductName,
-            //            Price = x.Product.Price,
-            //            SellingPrice = (decimal)x.Product.SellingPrice,
-            //            Pictures = (List<string>)x.Product.Pictures
-            //        })
-            //    .ToListAsync());
-
-            //return (await _context.WishList.Include(x => x.Product)
-            //    .Where(x => x.MemberId == myId)
-            //    .Select(x => new WishListGetModel
-            //    {
-            //        WishListId = x.WishListId,
-            //        ProductName = x.Product.ProductName,
-            //        Price = x.Product.Price,
-            //        SellingPrice = (decimal)x.Product.SellingPrice,
-            //        Pictures = x.Product.Pictures.Select(x => x.PicturePath).ToList(),
-            //        ProductId = x.ProductId,
-            //        MemberId = x.MemberId
-            //    }
-            //    )
-            //    .ToListAsync());
         }
 
         //TODO 移除收藏清單的商品 RemoveToWishlist
